Login SSO

+++ routes/api.php
@@ -19,55 +19,57 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
 });
 
 Route::post('login', 'AuthController@login');
+Route::post('/sit-login', 'AuthController@SITLogin');
+
 Route::group(['middleware' => ['checkAuth']], function () {
-    Route::middleware(['role.permission:admin'])->group(function () {
+    // Route::middleware(['role.permission:admin'])->group(function () {
         Route::get('/me', 'AuthController@me');
 
-        Route::get('company', 'CompanyController@get');
-        Route::post('company', 'CompanyController@create');
-        Route::put('company', 'CompanyController@update');
-        Route::delete('company', 'CompanyController@destroy');
+        Route::get('company', 'CompanyController@get')->middleware(['role.permission:access_company']);
+        Route::post('company', 'CompanyController@create')->middleware(['role.permission:create_company']);
+        Route::put('company', 'CompanyController@update')->middleware(['role.permission:update_company']);
+        Route::delete('company', 'CompanyController@destroy')->middleware(['role.permission:delete_company']);
 
         Route::prefix('academic-industry')->group(function () {
-            Route::get('job-positions', 'JobPositionController@get');
-            Route::get('announcement', 'AnnouncementController@get');
-            Route::get('announcements/{company_id}', 'AnnouncementController@getAnnouncementByCompanyId');
-            Route::get('announcements', 'AnnouncementController@getAnnouncements');
-            Route::post('announcement', 'AnnouncementController@create');
-            Route::put('announcement', 'AnnouncementController@update');
-            Route::delete('announcement', 'AnnouncementController@destroy');
+            Route::get('job-positions', 'JobPositionController@get')->middleware(['role.permission:access_academic_announcement']);
+            Route::get('announcement', 'AnnouncementController@get')->middleware(['role.permission:access_academic_announcement']);
+            Route::get('announcements/{company_id}', 'AnnouncementController@getAnnouncementByCompanyId')->middleware(['role.permission:access_academic_announcement']);
+            Route::get('announcements', 'AnnouncementController@getAnnouncements')->middleware(['role.permission:access_academic_announcement']);
+            Route::post('announcement', 'AnnouncementController@create')->middleware(['role.permission:create_academic_announcement']);
+            Route::put('announcement', 'AnnouncementController@update')->middleware(['role.permission:update_academic_announcement']);
+            Route::delete('announcement', 'AnnouncementController@destroy')->middleware(['role.permission:delete_academic_announcement']);
 
-            Route::get('applications', 'ApplicationController@get');
-            Route::get('application/{application_id}', 'ApplicationController@getApplicationById');
-            Route::post('application', 'ApplicationController@create');
-            Route::put('application', 'ApplicationController@update');
+            Route::get('applications', 'ApplicationController@get')->middleware(['role.permission:access_academic_application']);
+            Route::get('application/{application_id}', 'ApplicationController@getApplicationById')->middleware(['role.permission:access_academic_application']);
+            Route::post('application', 'ApplicationController@create')->middleware(['role.permission:create_academic_application']);
+            Route::put('application', 'ApplicationController@update')->middleware(['role.permission:update_academic_application']);
             Route::delete('application/{application_id}', 'ApplicationController@destroy');
         });
 
         // keep for dashboard feature
-        Route::get('companies', 'CompanyController@getCompanies');
+        Route::get('companies', 'CompanyController@getCompanies')->middleware(['role.permission:access_company']);
 
-        Route::get('users', 'UserController@get');
-        Route::get('user/{user_id}', 'UserController@getUserById');
-        Route::post('user', 'UserController@create');
-        Route::put('user', 'UserController@update');
-        Route::delete('user/{user_id}', 'UserController@destroy');
+        Route::get('users', 'UserController@get')->middleware(['role.permission:access_user']);
+        Route::get('user/{user_id}', 'UserController@getUserById')->middleware(['role.permission:access_user']);
+        Route::post('user', 'UserController@create')->middleware(['role.permission:create_user']);
+        Route::put('user', 'UserController@update')->middleware(['role.permission:update_user']);
+        Route::delete('user/{user_id}', 'UserController@destroy')->middleware(['role.permission:delete_user']);
 
         Route::get('roles', 'RoleController@get');
         Route::get('role-permissions', 'RoleController@getRolePermissions');
 
-        Route::get('histories', 'HistoryController@get');
+        Route::get('histories', 'HistoryController@get')->middleware(['role.permission:access_history']);
 
-        Route::get('banners', 'BannerController@get');
-        Route::get('banner', 'BannerController@getBannerById');
-        Route::post('banner', 'BannerController@create');
-        Route::delete('banner', 'BannerController@destroy');
+        Route::get('banners', 'BannerController@get')->middleware(['role.permission:access_academic_banner']);
+        Route::get('banner', 'BannerController@getBannerById')->middleware(['role.permission:access_academic_banner']);
+        Route::post('banner', 'BannerController@create')->middleware(['role.permission:create_academic_banner']);
+        Route::delete('banner', 'BannerController@destroy')->middleware(['role.permission:delete_academic_banner']);
 
-        Route::prefix('dashboard')->group(function () {
+        Route::group(['prefix' => 'dashboard', 'middleware' => ['role.permission:access_dashboard']], function () {
             Route::get('stats', 'DashboardController@getStats');
             Route::get('company-types', 'DashboardController@getCompanyTypes');
             Route::get('students/job-positions', 'DashboardController@getStudentJobPositions');
             Route::get('announcements/job-positions', 'DashboardController@getAnnouncementJobPositions');
         });
-    });
+    // });
 });
